requirements.txt
Django==4.1.6
djangorestframework==3.12.4
psycopg2-binary==2.8.6
